Move test into the right place
